pages: Updated cache sweeper

+++ lib/pages_core/cache_sweeper.rb
@@ -3,6 +3,15 @@ module PagesCore
 	class CacheSweeper
 		class << self
 
+			attr_accessor :enabled
+
+			def disable(&block)
+				old_value = self.enabled
+				self.enabled = false
+				yield
+				self.enabled = old_value
+			end
+
 			# Returns the default configuration.
 			def default_config
 				default = MethodedHash.new
@@ -34,30 +43,34 @@ module PagesCore
 
 			# Sweep all cached pages
 			def sweep!
-				cache_base_dir = self.config.cache_path
-				swept_files = []
-				if PagesCore.config(:domain_based_cache)
-					cache_dirs = Dir.entries(cache_base_dir).select{|d| !(d =~ /^\./) && File.directory?(File.join(cache_base_dir, d))}.map{|d| File.join(cache_base_dir, d)}
-				else
-					cache_dirs = [cache_base_dir]
-				end
-				cache_dirs.each do |cache_dir|
-					if File.exist?(cache_dir)
-						kill_patterns = self.config.patterns
-						paths = []
-						Find.find(cache_dir+"/") do |path|
-							Find.prune if path =~ Regexp.new("^#{cache_dir}/dynamic_image") # Ignore dynamic image
-							file = path.gsub(Regexp.new("^#{cache_dir}"), "")
-							kill_patterns.each do |p|
-								if file =~ p && File.exist?( path )
-									swept_files << path
-									`rm -rf #{path}`
+				if self.enabled
+					cache_base_dir = self.config.cache_path
+					swept_files = []
+					if PagesCore.config(:domain_based_cache)
+						cache_dirs = Dir.entries(cache_base_dir).select{|d| !(d =~ /^\./) && File.directory?(File.join(cache_base_dir, d))}.map{|d| File.join(cache_base_dir, d)}
+					else
+						cache_dirs = [cache_base_dir]
+					end
+					cache_dirs.each do |cache_dir|
+						if File.exist?(cache_dir)
+							kill_patterns = self.config.patterns
+							paths = []
+							Find.find(cache_dir+"/") do |path|
+								Find.prune if path =~ Regexp.new("^#{cache_dir}/dynamic_image") # Ignore dynamic image
+								file = path.gsub(Regexp.new("^#{cache_dir}"), "")
+								kill_patterns.each do |p|
+									if file =~ p && File.exist?( path )
+										swept_files << path
+										`rm -rf #{path}`
+									end
 								end
 							end
 						end
 					end
+					return swept_files
+				else
+					return []
 				end
-				return swept_files
 			end
 
 			# Sweep cached dynamic image
@@ -76,6 +89,8 @@ module PagesCore
 				end
 				[]
 			end
+
+			self.enabled ||= true
 		end
 	end
 end
